requirements.txt
# Core dependencies
Flask>=2.0.0
requests>=2.25.0
beautifulsoup4>=4.9.0
python-dotenv>=0.19.0

# Web scraping and parsing
lxml>=4.6.0

# Testing
pytest>=6.0.0
pytest-cov>=2.12.0

# Code quality  
flake8>=4.0.0

# Development utilities
click>=8.0.0